requirements.txt
numpy
time
glob
os
pillow